pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "py-solis-s3-logger"
description = ''
readme = "README.md"
requires-python = ">=3.7"
license = "MIT"
keywords = []
authors = [
  { name = "Torben Tretau", email = "torben@tretau.net" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "click",
  "httpx",
  "httpx-auth",
  "paho-mqtt",
  "timed-count"
]
dynamic = ["version"]

[project.urls]
Documentation = "https://github.com/ttretau/py-solis-s3-logger#readme"
Issues = "https://github.com/ttretau/py-solis-s3-logger/issues"
Source = "https://github.com/ttretau/py-solis-s3-logger"

[project.scripts]
py-solis-s3-logger = "py_solis_s3_logger.cli:py_solis_s3_logger"

[tool.hatch.version]
path = "py_solis_s3_logger/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "pytest",
  "pytest-cov",
  "pytest-httpx"
]
[tool.hatch.envs.default.scripts]
cov = "pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=py_solis_s3_logger --cov=tests {args}"
no-cov = "cov --no-cov {args}"
[tool.hatch.envs.production.scripts]
publisher = "py-solis-s3-logger"
[[tool.hatch.envs.test.matrix]]
python = ["37", "38", "39", "310", "311"]

[tool.hatch.build.targets.sdist]
include = ["/py_solis_s3_logger"]

[tool.hatch.envs.production]
dev-mode = false
features = []
path = ".venv"

[tool.coverage.run]
branch = true
parallel = true
omit = [
  "py_solis_s3_logger/__about__.py",
]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
